+++ Fun.Blazor.Server/DIExtensions.fs
@@ -250,23 +250,30 @@ type FunBlazorServerExtensions =
         builder.AddEndpointFilter(FunBlazorEndpointFilter(defaultArg preventStreamingRendering false, defaultArg statusCode 200))
 
 
-    static member private MakeCreateAttrFn(ty: Type) =
+    static member private MakeCreateAttrFn(ty: Type, ?forCustomElement) =
+        let forCustomElement = defaultArg forCustomElement false
+
         let paramsInfo =
             ty.GetProperties(BindingFlags.Instance ||| BindingFlags.Public)
-            |> Seq.choose (fun p -> 
+            |> Seq.choose (fun p ->
                 let attr = p.GetCustomAttribute<ParameterAttribute>()
-                if isNull attr then None else Some(p.Name, p.PropertyType)
+                if isNull attr then
+                    None
+                else
+                    if forCustomElement && p.Name <> "FunBlazorDebugKey" && (p.Name |> Seq.exists Char.IsUpper) then
+                        failwith
+                            $"Parameter name ({p.Name}) for custom element component {ty.FullName} must be lowercase name with low dash for multiple words, like xxx or xxx_xxx"
+                    Some(p.Name, p.PropertyType)
             )
             |> Map.ofSeq
 
-        let parseValue (name: string) (v: string) (ty: Type) = 
-            if ty = typeof<string> then box v
+        let parseValue (name: string) (v: string) (ty: Type) =
+            if ty = typeof<string> then
+                box v
             else
-                let ty =
-                    if ty.Name.StartsWith("Nullable`") then ty.GetGenericArguments()[0]
-                    else ty
+                let ty = if ty.Name.StartsWith("Nullable`") then ty.GetGenericArguments()[0] else ty
                 ty.GetMethods(BindingFlags.Public ||| BindingFlags.Static)
-                |> Seq.tryFind (fun x -> 
+                |> Seq.tryFind (fun x ->
                     let ps = x.GetParameters()
                     x.Name = "Parse" && ps.Length = 1 && ps[0].ParameterType = typeof<string>
                 )
@@ -279,12 +286,12 @@ type FunBlazorServerExtensions =
             |> Seq.choose (fun p ->
                 let matchQuery () =
                     match ctx.Request.Query.TryGetValue p.Key with
-                    | true, v when v.Count > 0 -> Some (p.Key => parseValue p.Key (v.Item(v.Count - 1).ToString()) p.Value)
+                    | true, v when v.Count > 0 -> Some(p.Key => parseValue p.Key (v.Item(v.Count - 1).ToString()) p.Value)
                     | _ -> None
                 if ctx.Request.Method.Equals(HttpMethods.Post, StringComparison.OrdinalIgnoreCase) then
                     match ctx.Request.Form.TryGetValue(p.Key) with
-                    | true, v when v.Count > 0 -> Some (p.Key => parseValue p.Key (v.Item(v.Count - 1).ToString()) p.Value)
-                    | _ -> matchQuery()
+                    | true, v when v.Count > 0 -> Some(p.Key => parseValue p.Key (v.Item(v.Count - 1).ToString()) p.Value)
+                    | _ -> matchQuery ()
                 else
                     matchQuery ()
             )
@@ -304,22 +311,20 @@ type FunBlazorServerExtensions =
                 |}
         )
 
-        if not Utils.isRazorComponentsForSSRMapped then
-            Utils.isRazorComponentsForSSRMapped <- true
-            builder
-                .Map(
-                    "/fun-blazor-server-side-render-components/{componentType}",
-                    Func<_, _, _>(fun (componentType: string) (ctx: HttpContext) ->
-                        match Utils.razorComponentsForSSRTypes.Value.TryGetValue(componentType) with
-                        | true, comp -> html.blazor (comp.Type, attr = comp.CreateAttr ctx)
-                        | _ -> defaultArg notFoundNode html.none
-                    )
+        builder
+            .Map(
+                "/fun-blazor-server-side-render-components/{componentType}",
+                Func<_, _, _>(fun (componentType: string) (ctx: HttpContext) ->
+                    match Utils.razorComponentsForSSRTypes.Value.TryGetValue(componentType) with
+                    | true, comp -> html.blazor (comp.Type, attr = comp.CreateAttr ctx)
+                    | _ -> defaultArg notFoundNode html.none
                 )
-                .AddFunBlazor()
-            |> ignore
+            )
+            .AddFunBlazor()
+        |> ignore
 
     /// This will serve all blazor components (inherit from ComponentBase) in the target assembly for server side rendering
-    /// route pattern: /fun-blazor-server-side-render-components/{componentType}. 
+    /// route pattern: /fun-blazor-server-side-render-components/{componentType}.
     /// Value can be passed by query or form for component property, form will have higher priority. Only the last value will be take when found same keys.
     [<Extension>]
     static member MapRazorComponentsForSSR(builder: IEndpointRouteBuilder, assembly: Assembly, ?notFoundNode: NodeRenderFragment) =
@@ -339,25 +344,23 @@ type FunBlazorServerExtensions =
             Utils.funBlazorCustomElementsForSSRTypes.Value[ty.FullName] <-
                 {|
                     Type = ty
-                    CreateAttr = FunBlazorServerExtensions.MakeCreateAttrFn ty
+                    CreateAttr = FunBlazorServerExtensions.MakeCreateAttrFn(ty, forCustomElement = true)
                 |}
         )
 
-        if not Utils.isFunBlazorCustomElementsForSSRMapped then
-            Utils.isFunBlazorCustomElementsForSSRMapped <- true
-            builder
-                .Map(
-                    "/fun-blazor-custom-elements/{componentType}",
-                    Func<_, _, _>(fun (componentType: string) (ctx: HttpContext) ->
-                        match Utils.funBlazorCustomElementsForSSRTypes.Value.TryGetValue(componentType) with
-                        | true, comp ->
-                            let attrs = comp.CreateAttr ctx
-                            html.customElement (comp.Type, attrs = attrs)
-                        | _ -> defaultArg notFoundNode html.none
-                    )
+        builder
+            .Map(
+                "/fun-blazor-custom-elements/{componentType}",
+                Func<_, _, _>(fun (componentType: string) (ctx: HttpContext) ->
+                    match Utils.funBlazorCustomElementsForSSRTypes.Value.TryGetValue(componentType) with
+                    | true, comp ->
+                        let attrs = comp.CreateAttr ctx
+                        html.customElement (comp.Type, attrs = attrs)
+                    | _ -> defaultArg notFoundNode html.none
                 )
-                .AddFunBlazor()
-            |> ignore
+            )
+            .AddFunBlazor()
+        |> ignore
 
     /// This will serve all components which is marked as FunBlazorCustomElementAttribute in the target assembly for server side rendering with custom element enabled,
     /// You should use it with: services.AddServerSideBlazor(fun options -> options.RootComponents.RegisterCustomElementForFunBlazor<YourComponent>()),